adding fixed IPs for Mesos like k8s annotations
